feat(breakpoints store): setup basic store
